requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
Django==6.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
IMAPClient==3.0.1
jiter==0.12.0
openai==2.15.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-telegram-bot==20.3
requests==2.32.5
secure-smtplib==0.1.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
